micro_zip: improve compression a bit, remove comments
Now we save the compressed version of a file in the result ZIP file if
it that saves at least one byte, instead of using a more space wasting
threshold, because resource pack loading performance doesn't seem to be
affected that much.

Also, it's very likely that we won't ever need to store folders in a ZIP
file, so remove comments that made reference to them to make the source
code cleaner.