Accept comments in input JSON files
Minecraft uses the lenient parser of the Gson library to parse JSON
files. This allows Minecraft to read JSON files that do not conform to
the JSON spec, because that lenient mode extends the JSON grammar to
allow comments and a bit less verbosity.

However, the JSON parser used by PackSquash follows the JSON spec to the
letter, which means that some common JSON extensions are not allowed.
This can generate frustration to end users and make them wonder why
PackSquash doesn't accept JSON files that Minecraft does.

As a compromise between user expectations, maturity and performance of
Rust libraries and educating users to write JSON that will work no
matter what the internals of Minecraft are, we now strip comments from
JSON files before parsing them. This does not affect the output (so the
JSON files that PackSquash puts in the resulting ZIP file do not contain
comments). Comments were deemed important for users because PackSquash
is commonly used in a CI-like manner, where the original files of a
repository are all processed at once, and comments are useful for
documentation and team work in this kind of environment.