Use half of available memory for spooling buffer size
This lets PackSquash use half of the available main memory that the
operating system reports as a buffer to store parts of the in-progress
ZIP file that is being generated. Previously, only up to 64 MiB were
used for this purpose.

This commit also exposes a setting for the user to manually override
this computed size.

In any case, if the file grows to a size bigger than this, it will be
flushed to disk and all operations will then be carried on that file.