audio_file: maybe finally fix test deadlocks
Investigating the causes of this deadlock led to the conclusion that the
most probable cause is that signalling EOS without passing GStreamer any
input buffer triggers some kind of undefined behavior that may or may
not work.

Fix that by not processing empty files with GStreamer altogether, and
also add a separate test to ensure that GStreamer works okay with non
empty, but still invalid files.