audio_file: do not require ReadAsync to be Send. Get rid of aux task
This commit lets the user of the returned stream to be the one that
drives forwarding input data to the appsink, instead of an auxiliary
task. This is done by polling another stream that polls the forwarding
future to completion, in addition to the rest of streams already used
internally.