Handle and ignore BOM for text files
Some users reported via Discord that some text files could not be parsed
correctly due to mysterious reasons, even though they looked valid. It
turned out that the cause of these failures was the presence of a BOM at
the beginning of the file, which is read as just another Unicode
character and breaks most parsers, because they usually don't expect it.

Fix the situation by ignoring the BOM, as byte order is not relevant for
UTF-8 and just adds to file size. Also add test cases to ensure this
works as expected.