system_id: tidy up and make ID selection much more smarter
Now it uses all the available methods to gather a system ID for a host,
and chooses among them according to their entropy, volatileness and
hardcoded priority. In general, the method that offers the highest
entropy will be used, unless the ID it provides is volatile and not
substantially better in entropy than a non-volatile method, and in the
case that a tie happens the entropy with highest hardcoded priority will
be preferred.

Of course, the system ID provided in the relevant environment variable
will take preference over all of this, because we trust the user and/or
system administrator to put a good ID and not let the environment be
controlled by malicious actors.