Handle unavailability of file creation and modif. times properly
Previously, a hard error would occur if the file creation or
modification time is not available for the filesystem, which is not
desireable. Now PackSquash uses both times only if available, and if the
modification time is not available, it will fall back as if no time
information at all was available, because creation times are very
unreliable: they are not updated when a file is changed.