Lazily open pack files, refactor properties error type
By lazily opening pack files we can do so in a multithreaded fashion,
and also save file handles and syscalls if the pack file constructor
determines that the file should be skipped.

Also refactor the properties error type, so we are shielded against
changes in PropertiesError, and maybe we can implement Send or Sync if
needed.