Possible proper fix for Windows executable metadata missing
It turns out that winres doesn't actually work when a package contains
both binary and library crates, due to limitations on how the build
script Cargo keys responsible for linking are interpreted. So say
goodbye to winres, and reinvent the wheel by adding the resource data
after the executable is built and not during linking, because the wheels
we have are punctured for our use case.