Stop reading and using file creation time
The file modification time is set to the creation time when a file is
just created, so it's a reliable way to get when the file contents were
modified for the last time, assuming proper filesystem and kernel
configurations. The creation time is less reliable (for instance, it is
not available on NTFS filesystems mounted with ntfs-3g under Linux), and
it can't be faked (which may be useful for advanced users, because they
might know better when a file was modified or not, and they may desire
that PackSquash honors file metadata that may be set with tools like
`touch`). Also, if the file metadata reports that a file was created
after it was modified, we know that some date does not make sense, but
favoring one over another is not really a proper fix either (was the
file actually modified for the last time that far back, and is the newer
creation time just a result of copying the file over from another
filesystem, for example?). So let's dumb down the code, making it
simpler and behave even better in some advanced scenarios.