Fix handling of packs with lots of file, adding a new option
Huge packs may contain lots of files, and because we spawn a task for
each file as fast as we read the directory entries from the filesystem,
if we don't process the files as quick as we open them we can exhaust
system resources.

Avoid that by limiting the number of concurrent tasks according to the
file limit, which should be considerably higher than the number of
threads, and not limit performance in the real world. Expose a
configuration option just in case someone wants to tweak that anyway.