squash_zip: support empty files
Empty files triggered a debug assertion panic in the Zopfli compressor.
This shouldn't happen in release builds, but there is little point in
doing Zopfli compression over nothing anyway, so just skip it for empty
files.

Add some tests to make sure that adding empty or very small files to a
ZIP file works as expected.