Stop processing a pack if an error occurs
Because an error may occur almost anywhere while processing a pack for
very different reasons, it is not possible to guarantee, in general,
that after an error occurs the user-expected invariants are held;
namely, that the generated output ZIP file is not corrupt and does not
miss necessary files. Let's make PackSquash more robust even if errors
happen by failing-fast, and returning an appropriate status code to the
operating system. Paired with a performance-focused options file, this
behavior allows using PackSquash as a pack validator somewhat
effectively, too.

The downside of this change is that sometimes certain errors may be
tolerable, and fixing an already existing resource pack that does not
validate may be time consuming, but I think it's fair that PackSquash is
a bit opinionated and rejects junk that just happens to work okay-ish in
Minecraft.