Experimental: cross-compile for Linux ARM64 CPUs
These builds are useful for those that want to run PackSquash on a
modern ARM CPU, like that in the Raspberry Pi 3 onwards, under a
GNU/Linux environment. Unit tests are not run for this target and the
builds have not actually been tested on an ARM device yet, though.