Addendum to e26b18e4e2c0d9ed26bc939e4623309ae41f1eea
This is a more proper and complete fix for the unintended behavior
change documented in the referenced commit, that descends into root
directories that should be ignored if they were not root. In general, we
want to always descend into the root directory the user provides, no
matter if it is a hidden or system directory or not, and then filter the
contents of that directory as usual.