Add stick parity bit to sanitized Squash Times
Although this stick parity bit made it necessary to reduce the timestamp
precision from quarter-seconds to half-seconds in order to keep the same
range of representable dates, it is a worthy tradeoff, because it allows
to detect whether a Squash Time was decrypted with a wrong key or
modified with 50% probability. Because practical packs have a
considerable number of files and the probability that such errors in a
systematic way get undetected are (1/2)^file count, this is a pretty
good, and in lots of cases accurate, way to detect whether such errors
happen.