Update dependencies. Remove unneeded dev dependencies
The update of the `time` crate disables some features by default now,
which helps bringing down the amount of transitive dependencies for
PackSquash. To compensate for the lesser features, we now use other
method to get a build date in the build script that provides equivalent
results in similar amounts of code.

The `fpe` crate was also updated with several changes, including those
that we did on our fork and submitted upstream via a PR, so stop
depending on our fork.