Add image_data_compression_iterations and skip_alpha_optimizations options
These options are meant to both substantially improve PNG optimization
performance, in spite of further potential OxiPNG optimizations, and
suitability of the optimization process to keep steganographic data
intact.

The model used in SquashZip to predict an appropriate number of Zopfli
compression iterations was moved to its own reusable struct, and it is
now used by the PNG file optimization logic too, albeit with different
parameters. If the Zopfli compression iterations drop to zero, a more
traditional and fast Zlib library is used, which greatly alleviates the
performance problems in bigger images.

Preliminary tests on a small resource pack with small textures showed
that compression is not affected when using the default values for these
options, which is a good thing.