GitHub CI: temporarily downgrade Linux runners to Ubuntu 18.04
The changes introduced in a1a02cfda35c9203b31adc614a345e468b7e00de did
not have the intended effect of getting rid of too new glibc symbols in
the generated binaries. Because any further workarounds are messy, let's
use shotgun surgery and downgrade the whole distro to guarantee that
nothing is too new.

I guess it is too complicated for glibc maintainers to add some
user-friendly options to choose the versions of the symbols to link to,
instead of having to resort to the unsupported static linking or symver
madness...