pack_file: stop using file size for framing decisions
Under pathological concurrent pack file modification scenarios, assuming
that the file size is constant can lead to non-termination if the file
size is increased since PackSquash first opened the file and retrieved
its metadata, because the framing decoders used to process JSON, PNG,
properties and shader files knew that the whole file was read to memory
by comparing the bytes read to the file size, not yielding any data
until this condition was met.

So, instead of relying on the file size to know whether it is a good
time to process all the data, implement the `decode_eof` method
suitably, which is guaranteed to be called when EOF is reached, with all
the unframed bytes read so far. This way we handle these pathological
cases gracefully on our part.

An interesting state machine diagram with all the states a Tokio
FramedRead can be can be found at https://github.com/tokio-rs/tokio/blob/08ed41f339e345286f26bdbc6c67bf9f2975ed07/tokio-util/src/codec/framed_impl.rs#L123-L151