Cargo.toml: set Debian package revision to zero
We don't have a proper v0.3.0-rc.1 Debian package to make available for
download, neither are ready to release the actual final v0.3.0 version.
However, we want to let users install something close to v0.3.0 via the
APT repository still, and make sure that they get updated to the final
v0.3.0 release when it is ready. Let's use an initial revision of 0 to
properly re-release v0.3.0 when the time comes.