GitHub CI: set CONST_RANDOM_SEED environment variable to a secret
This secret was introduced a month ago, along with the caching feature,
but it was not actually used in CI runs because it is necessary to
explicitly declare an environment variable with it.

Paired with the recently introduced stick parity bit in Squash Times,
this change should make reusing cached ZIP files accross PackSquash
builds much more effective and useful, even though there are still no
guarantees that it won't change between builds.