Refactor RelativePath so it validates path length too
We previously used raw strings at some points and used documentation to
enforce the contract that ZIP file names should be 65535 bytes long or
less, when we could let RelativePath also handle that size check, in
addition to the UTF-8 validity check.

As a nice side effect of this change, too long file paths are now
discarded very early in the pack file iteration process, which helps
avoiding useless computation in that unlikely case.