Add MTR3 support. Refactor JSON file structs and introduce asset types
This commit addresses issue #42. Specific debloating support was added
to .bbmodel files, as closer inspection of how the file format was
parsed by the mod revealed that a lot of data stored in those files was
not read at all.

While at it, the JSON pack file struct was refactored to use asset types
internally, which allows changing how each JSON file is optimized
according to the asset type it contains. Currently it is used to avoid
having to store the file extension in the heap, do asset-specific
debloating, and validating that most asset types are JSON objects, and
not any other kind of JSON value. In the future, this will provide the
plumbing needed to actually parse these asset types and achieve greater
optimization and/or validation. It will also be useful to apply
asset-specific quirks. Other pack file types will be refactored to use
asset types too.

Tests were updated accordingly, and the behavior of JSON file
optimization changed in some ways. First, due to the introduction of
asset types, now only JSON files that are within a Minecraft pack folder
structure will be added. Secondly, when debloating is done, now it is
shown in the optimization description.