Revert "Update to Rust 2021 edition"
This reverts commit 316ba54aec05e3c3631c45057f202e5af4e6713c.

Bisecting has shown that this commit caused PackSquash to not deal
properly with packs that contain a lot of files, and maybe caused some
additional slowdowns, maybe depending on the platform.

This commit is apparently inocous, though. Further investigation is
needed. Maybe some macro in third party code stopped working as
expected?