Show more user friendly error message if the output file path is a dir
This also avoids potentially doing a lot of useless work only for it to
go to waste due to an usage error.

The added check is only meant to improve user experience under common
usage failures. PackSquash should work correctly even if the output file
path somehow becomes a directory during its execution. In fact, the
output file path will be considered valid even if its parent directory
does not exist, and other similar failure scenarios. An attempt to avert
these is not made for the following reasons:

- The most proper way to deal with these situations is eagerly creating
  the output file, not at the end of the process. However, doing so has
  the bad side effect of nuking the already existing output file if
  something goes wrong (this can still happen if some I/O error occurs
  after opening the output file, while writing to it, but that is less
  common than the whole process going wrong because there is an error
  with the pack). I think it's better to risk wasting time than
  increasing the chance of unexpected data loss, even if it is the
  user's fault.
- Trying to anticipate other potential, less egregious failures in a
  portable way is fundamentally flawed. Anything can go wrong at any
  time. I think this hits the diminishing returns point, and further
  anticipations would cost too much resources or complexity compared to
  the benefits provided. For example, checking that the parent directory
  exists implies canonicalizing the path, which requires a buffer
  allocation (because the path can be relative, and getting the parent
  of relative paths is ugly), additional system calls and failure
  handling.