Open output ZIP file earlier. Reduce number of locks in squash_zip
In 233470d64bd9efc88706f5438d7afea65e66928c I said that opening the
output ZIP file earlier had the nasty side effect of deleting its
contents. However, after thinking about it more, I realized that does
not have to be the case: it is possible to open a file for writing,
creating it if necessary, without truncating it. Therefore, there are no
real reasons to not try opening that file sooner and let the OS catch
some errors in the process. In theory, this change could change
PackSquash behavior when the output ZIP file is a non-regular file, as
truncating non-regular files is either not defined or not specified.
However, I think that the additional user-friendliness is worth any
changes in this regard.

While at it, move fields that were locked together in SquashZip methods
to a common struct, so we only have to acquire a single lock to mutate
any of those fields, instead of one lock per field.