Revert attempts to fix ARM64 CI builds
It looks like the package lists are broken somehow. It'd make sense for
them to break from time to time, as Ubuntu package ports for ARM64 are
a community effort, and don't have the same degree of support as x64,
for example. I don't think that there is anything wrong with the
workflow either, so let's pretend that this issue does not exist and
move on, as it will likely be noticed by someone else who is in a better
position to do something about it.