Run Linux CI workflows in Debian containers
The GitHub-provided environments are based on Ubuntu, and have custom
package lists provided by Microsoft. Ubuntu in itself is not that good
for cross-compilation of native code because of its split and unofficial
repositories for multiarch support of even fairly common architectures,
such as arm64. And the custom package lists were not playing nice at all
with the upstream ports repository for the GStreamer packages we need,
because the upstream ports repository references versions of packages
that are not in our lists, generating cryptic error messages for apt.
And Ubuntu, being based on Debian, introduces yet another link in the
chain to upstream code, which complicates debugging some issues.
Finally, Debian has a much stricter stable release policy, which for our
purposes means that the latest stable Debian release ships with a glibc
version that is old enough to be already used by any other distros, but
still not obsolete, which makes the final executables more portable.

For all the reasons stated above, run the Linux builds inside a Docker
container that provides a Debian Bullseye environment. This comes at a
performance cost, but I think the benefits are worth it.