CI: run cargo deny to lint dependencies
This tool lints dependencies based on their security advisories, license
compatibility with AGPLv3 (in our case), maintainance status, and more
similar goodies.

cargo deny also discovered that the imagequant dependency did not have a
proper SPDX license identifier. A PR was submitted and merged to fix
that, so update the dependencies too.