CI: add experimental Linux musl builds
musl is an alternative system C library implementation that is mainly
used in Docker containers based on Alpine Linux, embedded systems and
SBCs. Let's try to support it. The lack of nice cross-compilation and
multiarch support by the distros for this task made it more complicated
than it should be.