Add build feature to disable GStreamer. Improve rustdoc comments
GStreamer is a nice library for multimedia processing, but its plugin
architecture and C implementation makes it a pain in the ass when
attempting to build executables for less common targets. It also can be
a bit burdensome to install. However, there are no real alternatives to
it, and it does its job well nevertheless.

To address this, add a build feature to disable the GStreamer
dependency. This is intended for advanced users that might find
giving up the audio transcoding, channel mixing and pitch shifting
features worth getting rid of the dependency when building PackSquash
from source. For the time being, the official PackSquash builds will
still use GStreamer as before.