Scaffold web UI package skeleton files
I've decided to go with webview and an embedded web server for the task.
A web UI is as cross-platform as it gets and also a good opportunity to
expose an HTTP API other applications can consume, or have a beefy
central computer that people can connect to in order to easily optimize
packs via a nice web interface. Resource consumption with respect to
other cross-platform alternatives is not that bad, and overall I think
this decision will make for a better UX for both developers and users.

Vue.js will be used for the frontend, which makes it a breeze to develop
SPAs that have separated view, styling and logic components. A SPA is
interesting here to minimize load on the pack optimization server and
provide a more native GUI feel.

When built, this package currently opens a window showing a Hello World
Vue.js app that was generated with the Vue CLI.

Things left to do:

- Actually design and implement the whole backend and frontend using the
  chosen tools
- Evaluate alternatives to actix-web for an embedded web server. Maybe
  another one is more lightweight, can reuse the Tokio runtime we
  already depend on, etc.
- Think on Windows and macOS support
- Integrate with the CI workflow. Build Debian packages and all that
  stuff, making sure its dependencies are correct
- Add tests, at least for the HTTP API offered by the backend
- Document properly somewhere