audio_file: do not upsample if the source sample rate is lower
PackSquash used to resample the audio file to whatever sampling
frequency was set in the audio file options, or the default one if no
options were set. This usually works well, but in the less likely case
that the source audio sampling frequency is less than the set sampling
frequency, this behavior effectively generates new samples that do not
improve quality and bloat the result file sizes.

Tweak that to resample to the lowest frequency among the source and set
frequencies, so that no useless samples are introduced if the source
audio file just does not contain them.