feat: add `always_allow_comments` option for JSON files
This option allows to instruct PackSquash to reject JSON comments in
files that do not have a specific extension that signals that they might
contain comments. This processing mode is faster and more
standards-compliant, but the default behavior of allowing and stripping
comments is kept because it is more ergonomic for pack authors.

The option is a file-specific option to allow choosing what files should
be fully standards-compliant in a fine-grained way. Keep in mind that
wildcard patterns can be used to match a lot of files at once.