refactor: handle failure to read back previous ZIP ourselves
When a previously generated ZIP file can't be reused, the user can't do
much about it except undoing whatever change made the ZIP file unusable,
or not using it to optimize this run. Undoing the change might be the
best course of action in some cases, but that requires user
intervention, and I don't expect that most people understands the reuse
feature enough to do the right thing (neither should they be expected
to, as a design goal of this feature was to make it as stateless and
transparent to the end user as possible).

To improve on this situation, let's make PackSquash ignore the previous
ZIP file automatically if something goes wrong while reading it. This
behavior is more amenable to scripts, most end users, and makes
PackSquash executions more "atomic" if for some reason it fails while
generating the ZIP file, leaving it in a bad state. To make advanced
users happy and, in general, let everyone know that the file could not
be reused, a warning with the details of what previously was an error is
emitted.