revert: stop opening output ZIP file earlier
While doing so has the advantage that I/O errors that may occur creating
the file are catched promptly, so the effort we spent optimizing the
pack does not go to waste, doing so in a technically nice way is not
possible. By creating it so soon PackSquash might leave a lingering
output ZIP file behind when it finishes unsuccessfully. It's possible to
anticipate some causes of termination and delete that file if they are
detected, but doing so is OS-dependent and still not immune to things
like abort()'ing.

To catch file creation I/O errors promptly without exposing ourselves to
the possibility of leaving an incomplete output ZIP file behind in any
case, which may confuse users and programs alike (in addition to feeling
"untidy" in my opinion), we would need to create an anonymous,
OS-managed temporary file in the directory where the resulting file
would be. This is feasible on Linux, with the O_TMPFILE flag and linkat.
However, not all Linux filesystems support such files, Windows can't
create anonymous temporary files, and other Unix systems don't support
O_TMPFILE at all. Therefore, such a concept is not portable, and the last
thing we want is for PackSquash to have different behavior in this
regard on different operating systems.

Therefore, all the possible ways to tackle this problem are bad in some
way. The previous way of opening the output ZIP file is the least bad,
because while it might waste user time (which is important), at least it
is technically more consistent, portable and simpler to maintain.

Note that it's still possible to leave lingering ZIP files if some error
occurs while populating its contents, but this is rarer than with the
prompt output file opening behavior.

Refs: 9fb10e6e9ece3b050823a6cee01174fd51ccfdb7