fix: do not allow files with duplicate names in ZIP files
Several files with the same name might be added to a ZIP file due to the
extension normalization feature. v0.2.1 prevented this from happening,
but while developing v0.3.0 I realized that the ZIP file format allows
duplicate files. This is technically true, but these duplicates do not
make sense, as at best one of them will be ignored by reader software,
so putting two can only cause trouble and less space efficiency.

PackSquash now will error out if two files are added to the ZIP file
with the same name. This is expected to be pretty rare, so the check is
done in the least intrusive way, without reording operations so that
it's catched ASAP.