fix(squash_zip): honor spool buffer size for scratch files properly
Previously the file size hint was used to establish the size threshold,
which is incorrect: we want to use it to set a capacity only. The
capacity will be bounded by the size threshold.

This should reduce memory usage if the file size is smaller than half
the buffer size, too.