feat: custom asset inclusion support
A pack author may desire to include unnecessary files to a pack for
several reasons, the most prevalent one being adding credits information
in a separate file. Up until now, PackSquash didn't provide any means to
include such files in the ZIP file it generates, and editing them
afterwards is not feasible if extraction protection is enabled.

Change that by adding a new type of file specific options,
CustomFileOptions, to mark files matched by them as custom assets that
should be included in the generated file without any changes.

Note that this feature is not an excuse for PackSquash to not properly
support pack assets used by Minecraft or some mod. If you use this
feature to work around that lack of support, please file an issue about
it or get in touch.

Resolves #40.