feat: introduce asset types masks, and generate them from pack metadata
Asset types masks are just sets of asset types that PackSquash can
match. This set can be initialized according to the allowed mods,
options like `skip_pack_icon` and the range of target Minecraft versions
according to the pack metadata file, and it can be used to skip pattern
matching altogether for the asset types it excludes.

This feature is not only useful for discarding pack files that are only
read by previous Minecraft versions for packs that target newer
Minecraft versions: it also allowed to streamline the pack file skipping
code and make asset type matching potentially more efficient.

A new option, `automatic_asset_types_mask_detection`, is added to let
the user configure whether PackSquash should adjust the asset types mask
it's using according to the pack metadata or not.