fix(properties_file): always strip BOM
Previously, the initial BOM character would not be stripped from the
file, which could cause unexpected errors and OptiFine to not read the
file correctly, when not minifying. Moreover, the BOM would add 3
troublesome bytes.

To squeeze some savings while fixing the effects of software that
insists on adding byte order marks, let's strip the BOM in that case
too. Some minor refactors were done while at it.