feat: add legacy language file support
These files were used by Minecraft 1.12.2 and older, but ignored by
PackSquash. For some reason, the Minecraft Wiki editors decided to
remove every reference to these from the resource pack article, so when
scraping it to make the assets list I've missed them. Thanks to some
Discord user I can't recall right now for pointing this out.

The files are validated for duplicate keys and missing separators, even
though the game is just lenient and allows this nonsense without
warnings for some reason. The UTF-8 BOM, if present, is removed, to save
space and make end users happy that their file works with Minecraft even
if their text editor insists on adding a BOM. This BOM handling behavior
has my "better than Mojang™" award, although a file-specific option to
stop it is added, to handle the extremely rare case that the BOM is
actually expected to be part of some language string key.