bench: add automated benchmark suite based on Criterion.rs
Some PackSquash users contacted me about some tricky details of its
performance characteristics. I try to write code that is as efficient as
possible, and Rust helps with that, but performance is a complicated
topic. Up until now, performance was informally measured by my own
educated guesses and some manual test runs here and there, where I payed
attention to the total run time and resource consumption. However, we've
reached a point where any performance improvement is likely to be not
trivial, and my own careful testing may just not cut it. Moreover, not
having hard baselines to base decisions on is, in general, a bad thing.

To properly tackle those problems, let's introduce automated
benchmarking to PackSquash, leveraging the Criterion.rs framework, which
does a bunch of clever statistics that make it easy to assess how some
changes impact the performance. The benchmarks can also be profiled, and
because they are designed to have a relatively short duration we can do
that with a high sampling frequency, which is desirable to catch small
inefficiencies here and there.

The benchmarks use a curated dataset of resource packs that were
contributed by myself and some PackSquash users, with their written
permission. The packs were chosen due to their diverse asset
distribution and perceived representativeness of PackSquash usage. To
keep the repository lightweight, they are stored externally in a Google
Drive folder. Helper scripts are committed to handle the download and
setup of the pack dataset with ease. Some packs of the dataset are
encrypted with GPG using a cryptographically-secure, random password, as
they are the intellectual property of several people, and in some cases
I was explicitly told not to make them public.

The benchmarks mainly measure the wall-time PackSquash takes to optimize
packs, but on Linux platforms the performance counters kernel subsystem
is used to gather insights on additional metrics that are correlated
with it, such as CPU instruction and context switch counts. The
performance counters can pinpoint the cause of a performance regression
or improvement and are less sensible to noise from, for example, other
running processes, but wall-time is actually what we want to optimize
(in addition to a sane usage of other resources, like I/O and memory).

Future commits may make the GitHub Actions CI run these benchmarks, to
help ensuring that PackSquash stays performant with each commit.

Special thanks to the PackSquash users that discussed performance,
including the Discord users Michael and jilchu. I'd also like to thank
my friend and classmate Víctor for writing the useful sample-pack.py
script and sharing his smart ideas and understanding of mathematical
principles to aid performance optimization.

Co-authored-by: victorlf4 <victorlevosofernandez@gmail.com>