ci: do not run twice for PRs within the repository
When a PR is opened to merge commits from one branch to another, the CI workflow is triggered twice because both the push and pull request events happen. Let's stop that by running jobs only if the triggering event was not a PR, or if it was a PR, it was not made in this repository.