ci(windows): downgrade GStreamer to the latest stable 1.18.6 version
The 1.18 series is still supported, and the 1.18.6 patch version was
released some days ago. The 1.20 series, of which the 1.19 versions were
an unstable preview, have more substantial changes that broke a bit some
packaging and generated some ugly, but benign warnings during Windows
tests.

Let's use the latest stable patch version of the 1.18 series, which was
known to work fine and not print these warnings.

Refs: 49fa1d105d6dd2c6083c876be48d5f32c9d20d18