feat(mtr3): support textures in mtr/custom_directory again
With the introduction of asset types and proper path matching instead of
just blindly optimizing files by broad type, we started skipping PNG
files that are outside of the vanilla-like assets/<namespace>/textures
folder. However, the example MTR3 resource pack, on which widely
deployed MTR3 resource packs are based on, decided to put custom
textures on a directory that does not conform to the more common
hierarchy: assets/mtr/custom_directory.

The mod can deal with textures being placed in either directory, and
it's my personal opinion that putting them following the vanilla-like
hierarchy looks more tidy. However, them being in a different hierarchy
also allows PackSquash to do more effective filtering of undesired mod
files when the MTR3 mod is not allowed in the options. To avoid breaking
most MTR3 packs and take advantage of this likely unintended design
choice, let's support MTR3 custom textures inside the custom_directory
folder again, but only if the MTR3 mod is allowed in the options.