ci: build AppImage for Linux x64
AppImage is a pretty neat distribution format for applications,
consisting of an executable image of a tiny root filesystem that
contains the application and the files it needs to run, including
libraries which cannot be assumed to be readily available in any sane
glibc Linux distro. When run, this executable image runs the application
it embeds transparently, just as if it was an executable with no
dependencies, passing arguments to it.

The AppImage ecosystem also includes desktop integration and automatic
update programs. While they are lacking when compared to native distro
package management tools, and AppImages are inefficient by design
because they can't reuse libraries that may alredy be installed in the
system, they are the closest thing to a "package once, run everywhere"
solution for Linux for applications that are dynamically linked and/or
may require certain files to be present in the system.

Sadly, building AppImages for arm64 requires compiling it from source,
and doing so is pretty complicated, and probably not worth it.