dist: replace linuxdeploy by appimage-builder
linuxdeploy did not offer the results we expected, as the generated
AppImages were missing critical components, such as GLib, that may not
be available in the headless systems we want to target. Besides that,
the generated AppImage still used the system glibc, which is good if we
limit ourselves to target recent desktop distros, but with AppImage we
don't care that much about wasting a MiB here or there: we want an
user-friendly, totally self-contained executable that feels like it was
statically linked, to maximize the environments it can run on with
little user trouble. Distro specific packages are better when available
anyway.