ci(linux): bump containers to Bullseye again
Now that we have a proper AppImage in place, we don't need to feel much
pity for those that run PackSquash on distros with old glibc: they just
can use the AppImage and be happy. Besides, the AppImage itself acts as
a gentle reminder that it may be a good idea to upgrade, providing a
migration path instead of a hard requirement for running official
builds.