fix(appimage/recipe): run tests and change Alpine image used
As expected, all the defined tests passed, with the only exception of
the Alpine test, because appimage-builder runs useradd before the actual
test command, and the base Alpine images do not have that command
installed.

Therefore, to fix the Alpine test it was enough to use a custom image
based on alpine:3 with a single layer to install the package, by running
`apk add --no-cache shadow`.