ci(macos): use upstream GStreamer distribution
Homebrew GStreamer 1.20 packages are missing the gstapp plugin, which provides core elements for PackSquash. Homebrew people think that we should ask what's going on upstream, but a quick look at the 1.20.1 source release shows that the gstapp plugin was not removed or something weird like that. So it looks like a build problem somewhere.

Let's see if the upstream .pkg works better. As a bonus, they look like universal binaries, so maybe they work better on Apple Silicon.